+++ example/src/auth0-configuration.js
@@ -1,0 +1,6 @@
+const config = {
+  clientId: 'u7KBoASmHOpIFqwp77tmGRHnUVaMupBj',
+  domain: 'brucke.auth0.com',
+};
+
+export default config;